+++ src/Features/ERDDAP/Platform/Forecasts/Page/index-configurable.stories.tsx
@@ -1,4 +1,5 @@
 import React from "react"
+import { select } from "@storybook/addon-knobs"
 
 import { UnitSystem } from "Features/Units/types"
 import { DataTimeSeries } from "Shared/timeSeries"
@@ -9,8 +10,7 @@ import { platform } from "stories/platform"
 
 export default {
   component: ForecastChart,
-  title: "ERDDAP|Forecast",
-  includeStories: [],
+  title: "ERDDAP|Forecast/Configurable",
 }
 
 const forecast_type = "wave_height"
@@ -34,6 +34,13 @@ data.push({
   timeSeries: forecast.readings,
   unit: forecast.source.units,
 })
+
+export const configurable = () => {
+  const options = [UnitSystem.english, UnitSystem.metric]
+  const unit = select("Unit System", options, options[0], "unit-system-0")
+
+  return <ForecastChart data={data} unit_system={unit} type={forecast_type} />
+}
 
 export const english = () => <ForecastChart data={data} unit_system={UnitSystem.english} type={forecast_type} />